+++ og-sdk-example/app/src/test/java/com/factual/ogsdkexample/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.factual.ogbroadcastreceiverexample;
+package com.factual.ogsdkexample;
 
 import org.junit.Test;
